% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_recipe}
\alias{add_recipe}
\title{Add a Recipe to Each Split Object}
\usage{
add_recipe(rset_obj, recipe, prep = TRUE, ...)
}
\arguments{
\item{rset_obj}{An object created by one of the resampling
functions in the package (e.g. \code{vfold_cv}).}

\item{recipe}{A recipe (which is not required to be
pre-prepared).}

\item{...}{Options to pass to \code{prep.recipe} such as
\code{verbose} or \code{retain}.}

\item{rep}{A logical; should the recipe be \code{prep}ed on
each of the analysis sets?}
}
\value{
The same \code{rset} object with a recipe in each
 element of the \code{splits} column.
}
\description{
\code{add_recipe} will add a pre-defined recipe to each
  \code{rsplit} element of the \code{rset} object.
}
