% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/along.R
\name{along}
\alias{along}
\alias{along.rset}
\alias{along.rset}
\title{Compute Across Resamples}
\usage{
along(x, ...)

\method{along}{rset}(x, ..., .elem = "splits", .unlist = FALSE)
}
\arguments{
\item{x}{An object.}

\item{...}{Arguments to pass to \code{\link[purrr]{map}} such as a function.}

\item{.elem}{A characterstring for which column of the \code{rset} object to compute over}

\item{.unlist}{A logical indicating if the results should be unlisted prior to returning the values.}
}
\value{
A list or vector of results the same length as \code{nrow(x$splits)}.
}
\description{
Compute Across Resamples
}
