% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apparent.R
\name{apparent}
\alias{apparent}
\title{Sampling for the Apparent Error Rate}
\usage{
apparent(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A tibble with a single row and classes \code{apparent},
\code{rset}, \code{tbl_df}, \code{tbl}, and \code{data.frame}. The
results include a column for the data split objects and one column
called \code{id} that has a character string with the resample identifier.
}
\description{
When building a model on a data set and re-predicting the same data, the
performance estimate from those predictions is often called the
"apparent" performance of the model. This estimate can be wildly
optimistic. "Apparent sampling" here means that the analysis and
assessment samples are the same. These resamples are sometimes used in
the analysis of bootstrap samples and should otherwise be
avoided like old sushi.
}
\examples{
apparent(mtcars)
}
