% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{balance_groups}
\alias{balance_groups}
\alias{balance_observations}
\alias{balance_prop}
\title{Balance group distribution across folds}
\usage{
balance_groups(...)

balance_observations(...)

balance_prop(prop = 3/4, ...)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{prop}{The proportion of data to be retained for modeling/analysis.}
}
\value{
A list with three elements: \code{fn}, the function which will actually
produce groups for the resampling method, \code{args}, a list of
any balancing-method specific arguments that \code{fn} requires, and \code{type},
indicating what balancing method is being used.
}
\description{
\code{\link[=balance_groups]{balance_groups()}} assigns groups to folds so that approximately equal
numbers of groups are in each fold.
}
