% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{bca_interval}
\alias{bca_interval}
\title{BCA Interval low-level}
\usage{
bca_interval(stats, splits, fn, args = list(), alpha = 0.05)
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
