% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{bca_interval_wrapper}
\alias{bca_interval_wrapper}
\title{BCA Interval wrapper}
\usage{
bca_interval_wrapper(stat_name, fn, args = list(), dat, alpha)
}
\description{
Help calculate bootstrap confidence intervals for a statistic of interest.
}
