% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Bootstrap Sampling}
\usage{
boot(data, times = 25, strata = NULL, apparent = FALSE, oob = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{times}{The number of bootstrap samples.}

\item{strata}{A variable that is used to conduct stratified sampling. When not \code{NULL}, each bootstrap sample is created within the stratification variable.}

\item{apparent}{A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the \code{summary} function that require the apparent error rate.}

\item{oob}{A logical. Should the out-of-bootstrap samples (aka "out-of-bag" aka "OOB") be retained? For traditional bootstrapping, these samples are not generally used.}

\item{...}{Not currently used.}
}
\value{
An object with classes \code{"boot"} and \code{"rset"}. The elements of the object include a tibble called \code{splits} that contains a column for the data split objects and a column called \code{id} that has a character string with the resample identifier.
}
\description{
Bootstrap Sampling
}
\examples{
boot(mtcars, times = 2)
boot(mtcars, times = 2, apparent = TRUE)
boot(mtcars, times = 2, oob = FALSE)
}
