% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Nonparametric Bootstrap Confidence Intervals}
\usage{
boot_ci(resamples_object, ..., method = c("percentile", "bca", "student-t"),
  alpha = 0.05)
}
\arguments{
\item{resamples_object}{An `rsplit` object created by the `bootstraps` function}

\item{method}{"percentile", "bca", or "student-t"}

\item{alpha}{level of significance}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
