% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfold.R
\name{check_v}
\alias{check_v}
\title{Check \code{v} for v-fold CV}
\usage{
check_v(v, max_v, rows = "rows", call = rlang::caller_env())
}
\arguments{
\item{v}{The user-supplied value for \code{v}}

\item{max_v}{The highest value of \code{v} possible}

\item{call}{The execution environment of the calling function}

\item{resample}{What is being resampled? Defaults to "rows".}
}
\description{
Check \code{v} for v-fold CV
}
\examples{
try(check_v(10, 5))
try(check_v(-2, 5))

}
\keyword{internal}
