% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\alias{rsplit_complement}
\title{Determine the Assessment Samples}
\usage{
complement(x, ...)

rsplit_complement(x, ...)
}
\arguments{
\item{x}{An \code{rsplit} object}

\item{...}{Not currently used}
}
\value{
A integer vector.
}
\description{
This method and function help find which data belong in the analysis and
assessment sets.
}
\details{
Given an \code{rsplit} object, \code{complement()} will determine which
of the data rows are contained in the assessment set. To save space,
many of the \code{rset} objects will not contain indices for the
assessment split.

\code{rsplit_complement()} handles the determination for sets for most resampling
methods. Unless the row indices for the assessment set are already
determined, this function selects all of the rows that are not in the
analysis set and returns those as the assessment set.
}
\examples{
set.seed(28432)
fold_rs <- vfold_cv(mtcars)
head(fold_rs$splits[[1]]$in_id)
fold_rs$splits[[1]]$out_id
complement(fold_rs$splits[[1]])
}
\seealso{
\code{\link[=populate]{populate()}}
}
