% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement_default}
\alias{complement_default}
\title{Determine the Default Assessment Samples}
\usage{
complement_default(x)
}
\arguments{
\item{x}{An \code{rsplit} object}
}
\value{
A integer vector.
}
\description{
Given many types of \code{rsplit} objects, \code{complement_default} will determine
which of the data rows are contained in the assessment set.
}
\keyword{internal}
