% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{default_complement}
\alias{default_complement}
\title{Get the indices of the analysis set from the assessment set}
\usage{
default_complement(ind, n)
}
\arguments{
\item{ind}{A vector of integers for which rows of data belong in the
assessment set.}

\item{n}{A single integer for the total number of rows in the data set.}
}
\value{
A named list of integer vectors.
}
\description{
Get the indices of the analysis set from the assessment set
}
\keyword{internal}
