% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{drinks}
\alias{drinks}
\title{Sample Time Series Data}
\description{
Sample Time Series Data
}
\details{
Drink sales. The exact name of the series from FRED is:
"Merchant Wholesalers, Except Manufacturers' Sales Branches and Offices
Sales: Nondurable Goods: Beer, Wine, and Distilled Alcoholic Beverages Sales"

These data are now in the \code{modeldata} package.
}
