% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fingerprint}
\alias{fingerprint}
\title{Create a cryptographical hash value for \code{rset} objects.}
\usage{
fingerprint(x, ...)
}
\arguments{
\item{x}{An \code{rset} object.}

\item{...}{Not currently used.}
}
\value{
A character string.
}
\description{
This function uses the distinct rows in the data set and the column(s) for the
resample identifier and the splits to produce a character string that can be
used to determine if another object shares the same splits.
}
\details{
The comparison is based on the unique contents of the \code{id} and \code{splits}
columns. Attributes are not used in the comparison.
}
\examples{
set.seed(1)
fingerprint(vfold_cv(mtcars))

set.seed(1)
fingerprint(vfold_cv(mtcars))

set.seed(2)
fingerprint(vfold_cv(mtcars))

set.seed(1)
fingerprint(vfold_cv(mtcars, repeats = 2))
}
