% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_split.R
\name{inner_split}
\alias{inner_split}
\alias{inner_split.mc_split}
\alias{inner_split.grouped_mc_split}
\alias{inner_split.vfold_split}
\alias{inner_split.group_vfold_split}
\alias{inner_split.clustering_split}
\alias{inner_split.apparent_split}
\title{Inner split of the analysis set for fitting a post-processor}
\usage{
inner_split(x, ...)

\method{inner_split}{mc_split}(x, split_args, ...)

\method{inner_split}{grouped_mc_split}(x, split_args, ...)

\method{inner_split}{vfold_split}(x, split_args, ...)

\method{inner_split}{group_vfold_split}(x, split_args, ...)

\method{inner_split}{clustering_split}(x, split_args, ...)

\method{inner_split}{apparent_split}(x, ...)
}
\arguments{
\item{x}{An \code{rsplit} object.}

\item{...}{Not currently used.}

\item{split_args}{A list of arguments to be used for the inner split.}
}
\value{
An \code{rsplit} object.
}
\description{
Inner split of the analysis set for fitting a post-processor
}
\keyword{internal}
