% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{int_bca}
\alias{int_bca}
\title{Bootstrap bias-corrected confidence intervals}
\usage{
int_bca(object, ..., fn, args = list(), alpha = 0.05)
}
\arguments{
\item{object}{bootstrap resamples created by the \code{bootstraps} function}

\item{...}{One or more unquoted expressions separated by commas
that select columns containing individual bootstrap estimates.
You can treat variable names like they are positions.}

\item{fn}{function to calculate statistic of interest}

\item{args}{list of arguments passed to \code{fn}}

\item{alpha}{Level of significance}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
