% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{int_t}
\alias{int_t}
\title{Bootstrap Student-t confidence intervals}
\usage{
int_t(object, ..., var_cols, alpha = 0.05)
}
\arguments{
\item{object}{Bootstrap resamples created by the \code{bootstraps} function}

\item{...}{One or more unquoted expressions separated by commas
that select columns containing individual bootstrap estimates.
You can treat variable names like they are positions.}

\item{var_cols}{One or more unquoted expressions separated by commas
that select columns containing individual bootstrap estimate
of \emph{variance} for the statistics. These selections can be
wrapped in \code{\link[dplyr:vars]{dplyr::vars()}}. It is assumed that the order of
these columns is in the same order as the statistics selected
by \code{...}.}

\item{alpha}{Level of significance}
}
\description{
Calculate bootstrap confidence interval with student-t method
}
