% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lables.R
\name{labels.rsplit}
\alias{labels.rsplit}
\title{Find Labels from rsplit Object}
\usage{
\method{labels}{rsplit}(object, ...)
}
\arguments{
\item{object}{An `rsplit` object}

\item{...}{Not currently used.}
}
\value{
A tibble.
}
\description{
Produce a tibble of identification variables so that single 
 splits can be linked to a particular resample.
}
\examples{
cv_splits <- vfold_cv(mtcars)
labels(cv_splits$splits[[1]])
}
