% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_cv}
\alias{loo_cv}
\title{Leave-One-Out Cross-Validation}
\usage{
loo_cv(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Not currently used.}
}
\value{
An object with classes \code{"loo_cv"} and \code{"rset"}. The elements of the object include a tibble called \code{splits} that contains a column for the data split objects and one column called \code{id} that has a character string with the resample identifier.
}
\description{
Leave-One-Out Cross-Validation
}
\examples{
loo_cv(mtcars)
}
