% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{new_rset}
\alias{new_rset}
\title{Constructor for new rset objects}
\usage{
new_rset(splits, ids, attrib = NULL, subclass = character())
}
\arguments{
\item{splits}{A list column of \code{rsplits} or a tibble with a single column
called "splits" with a list column of \code{rsplits}.}

\item{ids}{A character vector or a tibble with one or more columns that
begin with "id".}

\item{attrib}{An optional named list of attributes to add to the object.}

\item{subclass}{A character vector of subclasses to add.}
}
\value{
An \code{rset} object.
}
\description{
Constructor for new rset objects
}
\keyword{internal}
