% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{perc_all}
\alias{perc_all}
\title{Percentile wrapper for multiple statistics}
\usage{
perc_all(object, ..., alpha = 0.05)
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
