% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{perc_interval}
\alias{perc_interval}
\title{Bootstrap Confidence Intervals}
\usage{
perc_interval(stats, alpha = 0.05)
}
\arguments{
\item{alpha}{level of significance}

\item{stat}{A statistic of interest, a vector, from \code{rsplit} object created by the \code{bootstraps} function}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
