% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Permutations}
\usage{
permute(data, ..., times = 25)
}
\arguments{
\item{data}{A data frame.}

\item{...}{One or more selector functions to choose which
variables will be randomized within the data. Unquoted variable
names and \code{dplyr} selectors (e.g. \code{\link[tidyselect:starts_with]{tidyselect::starts_with()}},
\code{\link[tidyselect:ends_with]{tidyselect::ends_with()}}, \code{\link[tidyselect:contains]{tidyselect::contains()}},
\code{\link[tidyselect:matches]{tidyselect::matches()}}, \code{\link[tidyselect:num_range]{tidyselect::num_range()}},
\code{\link[tidyselect:everything]{tidyselect::everything()}}, and \code{\link[tidyselect:one_of]{tidyselect::one_of()}}, etc.)
can be used. At least one column should be specified.}

\item{times}{The number of permutations.}
}
\value{
An tibble with classes \code{permutation}, \code{rset}, \code{tbl_df},
\code{tbl}, and \code{data.frame}. The results include a column for the
data split objects and a column called \code{id} that has a character
string with the permutation identifier.
}
\description{
A permutation is a version of the data set where \emph{some} columns have been
arranged in a random order.
}
\details{
There is no notion of an assessment set for this
technique. \code{\link[=assessment]{assessment()}} will return the original data frame
with no rows.
}
\examples{
set.seed(75483)

permute(mtcars, mpg)

library(dplyr)
permute(iris, starts_with("Sepal"))
}
