% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{prepper}
\alias{prepper}
\title{Wrapper Function for Preparing Recipes}
\usage{
prepper(split_obj, recipe, ...)
}
\arguments{
\item{split_obj}{An `rplit` object}

\item{recipe}{An untrained `recipe` object.}

\item{...}{Arguments to pass to [recipes::prep()]
such as `verbose` or `retain`.}
}
\description{
When working with the \pkg{recipes} package, a simple recipe
 must be *prepared* using the [recipes::prep()]
 function first. When using recipies with \pkg{rsample} it
 is helpful to have a function that can prepare a recipe
 across a series of `split` objects that are produced
 in this package. `prepper` is a wrapper function
 around [recipes::prep()] that can be used to do
 this. See the vignette on "Recipes and rsample" for an
 example.
}
