% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lables.R
\name{pretty.vfold_cv}
\alias{pretty.vfold_cv}
\alias{pretty.loo_cv}
\alias{pretty.apparent}
\alias{pretty.rolling_origin}
\alias{pretty.mc_cv}
\alias{pretty.nested_cv}
\alias{pretty.bootstraps}
\alias{pretty.group_vfold_cv}
\title{Short Decriptions of rsets}
\usage{
\method{pretty}{vfold_cv}(x, ...)

\method{pretty}{loo_cv}(x, ...)

\method{pretty}{apparent}(x, ...)

\method{pretty}{rolling_origin}(x, ...)

\method{pretty}{mc_cv}(x, ...)

\method{pretty}{nested_cv}(x, ...)

\method{pretty}{bootstraps}(x, ...)

\method{pretty}{group_vfold_cv}(x, ...)
}
\arguments{
\item{x}{An `rset` object}

\item{...}{Not currently used.}
}
\value{
A character vector.
}
\description{
Produce a chracter vector of describing the resampling method.
}
\keyword{internal}
