% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reshuffle_rset}
\alias{reshuffle_rset}
\title{"Reshuffle" an rset to re-generate a new rset with the same parameters}
\usage{
reshuffle_rset(rset)
}
\arguments{
\item{rset}{The \code{rset} object to be reshuffled}
}
\value{
An rset of the same class as \code{rset}.
}
\description{
This function re-generates an rset object, using the same arguments used
to generate the original.
}
\examples{
set.seed(123)
(starting_splits <- group_vfold_cv(mtcars, cyl, v = 3))
reshuffle_rset(starting_splits)

}
