% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reverse_splits}
\alias{reverse_splits}
\title{Reverse the analysis and assessment sets}
\usage{
reverse_splits(x, ...)
}
\arguments{
\item{x}{An \code{rset} or \code{rsplit} object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
This functions "swaps" the analysis and assessment sets of either a single
\code{rsplit} or all \code{rsplits} in the \code{split} column of an \code{rset} object.
}
\examples{
set.seed(123)
starting_splits <- vfold_cv(mtcars, v = 3)
reverse_splits(starting_splits)
reverse_splits(starting_splits$splits[[1]])

}
