% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{rolling_origin}
\alias{rolling_origin}
\title{Rolling Origin Forecast Resampling}
\usage{
rolling_origin(data, initial = 5, assess = 1, cumulative = TRUE,
  skip = 0, ...)
}
\arguments{
\item{data}{A data frame.}

\item{initial}{The number of samples used for analysis/modeling in the initial resample.}

\item{assess}{The number of samples used for each assessment resample.}

\item{cumulative}{A logical. Should the analysis resample grow beyond the size specified by \code{initial} at each resample?.}

\item{skip}{A integer indicating how many (if any) resamples to skip to thin the total amount of data points int eh analysis resample.}

\item{...}{Not currently used.}
}
\value{
An object with classes \code{"rolling_origin"} and \code{"rset"}. The elements of the object include a tibble called \code{splits} that contains a column for the data split objects and a column called \code{id} that has a character string with the resample identifier.
}
\description{
Rolling Origin Forecast Resampling
}
\examples{
set.seed(1131)
ex_data <- data.frame(row = 1:20, some_var = rnorm(20))
dim(rolling_origin(ex_data))
dim(rolling_origin(ex_data, skip = 2))
dim(rolling_origin(ex_data, skip = 2, cumulative = FALSE))
}
