% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{rsample}
\alias{rsample}
\title{Formula based on https://goo.gl/7oWC9F}
\usage{
rsample(population = 1e+09, confidence_interval = 0.95,
  error_margin = 0.05, response_distribution = 0.5)
}
\arguments{
\item{population}{Size of the entire population from which to draw the sample. Defaults to an arbitrarily large population, giving a recommend sample size for the worst case scenario.}

\item{confidence_interval}{Confidence that the interval within +/- margin of error contains the true parameter. Smaller interval requires larger sample size.}

\item{error_margin}{Statistic expressing the amount of random sampling error in a survey's results. The larger the margin of error, the less confidence one should have that the poll's reported results are close to the "true" figures. Smaller margin of error requires larger sample size.}

\item{response_distribution}{Expected distribution of responses, with 50/50 being the "worst case". Known expected distribution decreases the sample size.}
}
\value{
The recommended sample size.
}
\description{
Formula based on https://goo.gl/7oWC9F
}
\examples{
    rsample()
    rsample(population = 15000, confidence_interval = .9, error_margin = .1)
    {
      population = c(100, 200, 30000, 400000)
      sample_size_95_5 = rsample(population)
      sample_size_90_10 = rsample(population, confidence_interval = .9, error_margin = .1)
      data.frame(population, sample_size_95_5, sample_size_90_10)
    }
    plot(rsample(seq(1, 30000)) ~ seq(1, 30000), type = "l")
}
