% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{rsample2caret}
\alias{rsample2caret}
\alias{caret2rsample}
\title{Convert Resampling Objects to Other Formats}
\usage{
rsample2caret(object, data = c("analysis", "assessment"))

caret2rsample(ctrl, data = NULL)
}
\arguments{
\item{object}{An `rset` object. Currently,
`nested_cv` is not supported.}

\item{data}{The data that was originally used to produce the
`ctrl` object.}

\item{ctrl}{An object produced by `trainControl` that has
had the `index` and `indexOut` elements populated by
integers. One method of getting this is to extract the
`control` objects from an object produced by `train`.}
}
\value{
`rsample2caret` returns a list that mimics the
 `index` and `indexOut` elements of a
 `trainControl` object. `caret2rsample` returns an
 `rset` object of the appropriate class.
}
\description{
These functions can convert resampling objects between
 \pkg{rsample} and \pkg{caret}.
}
