% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_cv}
\alias{spatial_cv}
\title{Spatial or Cluster Cross-Validation}
\usage{
spatial_cv(data, v = 10, repeats = 1, coords, ...)
}
\arguments{
\item{data}{A data frame.}

\item{v}{The number of partitions of the data set.}

\item{repeats}{The number of times to repeat the V-fold partitioning.}

\item{coords}{A vector of variable names, typically spatial coordinates,
to partition the data into disjointed sets via k-means clustering.}

\item{...}{Extra arguments passed on to \code{\link[stats:kmeans]{stats::kmeans()}}.}
}
\value{
A tibble with classes \code{spatial_cv}, \code{rset}, \code{tbl_df}, \code{tbl}, and
\code{data.frame}. The results include a column for the data split objects and
one or more identification variables. For a single repeat, there will be
one column called \code{id} that has a character string with the fold identifier.
For repeats, \code{id} is the repeat number and an additional column called \code{id2}
that contains the fold information (within repeat).
}
\description{
Spatial or cluster cross-validation splits the data into V groups of
disjointed sets using k-means clustering of some variables, typically
spatial coordinates. A resample of the analysis data consists of V-1 of the
folds/clusters while the assessment set contains the final fold/cluster. In
basic spatial cross-validation (i.e. no repeats), the number of resamples
is equal to V.
}
\details{
The variables in the \code{coords} argument are used for k-means clustering of
the data into disjointed sets, as outlined in Brenning (2012). These
clusters are used as the folds for cross-validation. Depending on how the
data are distributed spatially, there may not be an equal number of points
in each fold.

When more than one repeat is requested, the basic V-fold cross-validation
is conducted each time. For example, if three repeats are used with \code{v = 10}, there are a total of 30 splits which are three groups of 10 that are
generated separately.
}
\examples{
data(ames, package = "modeldata")
spatial_cv(ames, v = 5, coords = c(Latitude, Longitude))
spatial_cv(ames, v = 5, repeats = 2, coords = c(Latitude, Longitude))

}
\references{
A. Brenning, "Spatial cross-validation and bootstrap for the assessment of
prediction rules in remote sensing: The R package sperrorest," 2012 IEEE
International Geoscience and Remote Sensing Symposium, Munich, 2012,
pp. 5372-5375, doi: 10.1109/IGARSS.2012.6352393.
}
