% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{splits}
\alias{splits}
\title{Extract the resampling information}
\usage{
splits(x, .elem = NULL)
}
\arguments{
\item{x}{An `rset` object}

\item{.elem}{A character value or \code{NULL}. If \code{NULL}, then the entire tibble is returned. Values of the argument can be any column in the tibble.}
}
\value{
A tibble, vector, or list depending on the value of \code{.elem}.
}
\description{
The convenience function can be used to extract the data structure that contains the resampling splits or a specific element of that structure.
}
\examples{
 
set.seed(36522)
bt <- bootstraps(mtcars, times = 3)
splits(bt)
splits(bt, "id")
splits(bt, "splits")
}
