% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{student_t_all}
\alias{student_t_all}
\title{Student-t wrapper for multiple statistics}
\usage{
student_t_all(object, ..., var_cols, alpha = 0.05)
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
