% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{two_class_dat}
\alias{two_class_dat}
\title{Two Class Data}
\description{
Two Class Data
}
\details{
There are artifical data with two predictors (\code{A} and \code{B}) and
a factor outcome variable (\code{Class}).

These data are now in the \code{modeldata} package.
}
