% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_ptype2.rset}
\alias{vec_cast.rset}
\title{vctrs compatibility functions}
\usage{
\method{vec_ptype2}{rset}(x, y, ...)

\method{vec_cast}{rset}(x, to, ...)
}
\arguments{
\item{x, y}{Objects.}

\item{...}{Used to pass along error message information.}

\item{to}{Type to cast to.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow rset objects to
work with vctrs.
}
