% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attrition}
\alias{attrition}
\title{Job Attrition}
\source{
The IBM Watson Analytics Lab website \url{https://www.ibm.com/communities/analytics/watson-analytics-blog/hr-employee-attrition/}
}
\value{
\item{attrition}{a data frame}
}
\description{
Job Attrition
}
\details{
These data are from the IBM Watson Analytics Lab.
 The website describes the data with \dQuote{Uncover the
 factors that lead to employee attrition and explore important
 questions such as \sQuote{show me a breakdown of distance
 from home by job role and attrition} or \sQuote{compare
 average monthly income by education and attrition}. This is a
 fictional data set created by IBM data scientists.}. There
 are 1470 rows.
}
\examples{
data(attrition)
str(attrition)
}
\keyword{datasets}
