% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{bca_interval_wrapper}
\alias{bca_interval_wrapper}
\title{BCA Interval wrapper}
\usage{
bca_interval_wrapper(stat_name, fn, args, dat, alpha)
}
\arguments{
\item{stat_name}{statistic name}

\item{fn}{function to calculate statistic of interest}

\item{args}{list of arguments passed to \code{fn}}

\item{dat}{resamples}

\item{alpha}{level of significance}
}
\description{
BCA Interval wrapper
}
