% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_bca}
\alias{boot_ci_bca}
\title{BCa Method}
\usage{
boot_ci_bca(bt_resamples, stat, func, alpha = 0.05, data = NULL)
}
\arguments{
\item{bt_resamples}{An `rsplit` object created by the `bootstraps` function}

\item{stat}{A statistic of interest}

\item{func}{A function which when applied to data returns a vector containing the statistic(s) of interest.}

\item{alpha}{level of significance}

\item{data}{NULL}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
