% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_perc}
\alias{boot_ci_perc}
\title{Percentile Method}
\usage{
boot_ci_perc(bt_resamples, stat, alpha = 0.05, data = NULL)
}
\arguments{
\item{bt_resamples}{An `rsplit` object created by the `boostraps` function}

\item{stat}{A statistic of interest}
}
\description{
Calculate boostrap confidence intervals for a statistic of interest.
}
