% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_perc}
\alias{boot_ci_perc}
\title{Percentile Method}
\usage{
boot_ci_perc(bt_resamples, alpha = 0.05, stat, ...)
}
\arguments{
\item{bt_resamples}{An `rsplit` object created by the `bootstraps` function}

\item{alpha}{level of significance}

\item{stat}{A statistic of interest}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
