% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_t}
\alias{boot_ci_t}
\title{Bootstrap Confidence Intervals}
\usage{
boot_ci_t(bt_resamples, stat, stat_var, alpha = 0.05, data = NULL,
  theta_obs, var_obs)
}
\description{
Bootstrap Confidence Intervals
}
\details{
Calculate boostrap confidence intervals for a statistic of interest

 @export
}
