% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_t}
\alias{boot_ci_t}
\title{Bootstrap Confidence Intervals}
\usage{
boot_ci_t(bt_resamples, alpha, data = NULL)
}
\description{
Bootstrap Confidence Intervals
}
\details{
Calculate boostrap confidence intervals for a statistic of interest

 @importFrom dplyr mutate
 @importFrom stats sd
 @importFrom dplyr as_tibble
 @export
}
