% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci_t}
\alias{boot_ci_t}
\title{Student-T Bootstrap Confidence Intervals}
\usage{
boot_ci_t(bt_resamples, alpha = 0.05, stat, stat_var, ...)
}
\arguments{
\item{bt_resamples}{An `rsplit` object created by the `bootstraps` function}

\item{alpha}{level of significance}

\item{stat}{A statistic of interest}

\item{stat_var}{The variance of each bootstrap resample}
}
\description{
Calculate bootstrap confidence intervals for a statistic of interest.
}
