% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_split_args}
\alias{.get_split_args}
\title{Get the split arguments from an rset}
\usage{
.get_split_args(rset)
}
\arguments{
\item{rset}{An \code{rset} object.}
}
\value{
A list of arguments used to create the rset.
}
\description{
Get the split arguments from an rset
}
\keyword{internal}
