% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{fill}
\alias{fill}
\title{Add Assessment Indicies}
\usage{
fill(x, ...)
}
\arguments{
\item{x}{A \code{rsplit} and \code{rset} object.}

\item{...}{Not currently used}
}
\description{
\code{fill()} has been deprecated and will be removed in version \code{0.0.4}.
Please use \code{\link[=populate]{populate()}} instead.
}
