% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{fill}
\alias{fill}
\title{Add Assessment Indicies}
\usage{
fill(x, ...)
}
\arguments{
\item{x}{A \code{rsplit} and \code{rset} object.}

\item{...}{Not currently used}
}
\value{
An object of the same time with the integer indicies.
}
\description{
Many \code{rsplit} and \code{rset} objects do not contain indicators for
the assessment samples. \code{fill} can be used to populate the slot
for the appropriate indices.
}
\examples{
set.seed(28432)
fold_rs <- vfold_cv(mtcars)

fold_rs$splits[[1]]$out_id
complement(fold_rs$splits[[1]])

fill(fold_rs$splits[[1]])$out_id

fold_rs_all <- fill(fold_rs)
fold_rs_all$splits[[1]]$out_id
}
