% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{int_pctl}
\alias{int_pctl}
\title{Bootstrap percentile confidence intervals}
\usage{
int_pctl(object, ..., alpha = 0.05)
}
\arguments{
\item{object}{Bootstrap resamples created by the \code{bootstraps} function}

\item{...}{One or more unquoted expressions separated by commas
that select columns containing individual bootstrap estimates.
You can treat variable names like they are positions.}

\item{alpha}{Level of significance}
}
\description{
Calculate bootstrap confidence interval with percentile method
}
