% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_cv}
\alias{loo_cv}
\title{Leave-One-Out Cross-Validation}
\usage{
loo_cv(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Not currently used.}
}
\value{
An tibble with classes \class{loo_cv}, \class{rset}, \class{tbl_df}, \class{tbl}, and \class{data.frame}. The results include a column for the data split objects and one column called \code{id} that has a character string with the resample identifier.
}
\description{
Leave-one-out (LOO) cross-validation uses one data point in the original set as the assessment data and all other data points as the analysis set. A LOO resampling set has ass many resamples as rows in the original data set.
}
\examples{
loo_cv(mtcars)
}
