% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_splits}
\alias{make_splits}
\title{Constructor for split objects}
\usage{
make_splits(ind, data, class = NULL)
}
\arguments{
\item{ind}{A list of integers with names "analysis" and "assessment".}

\item{data}{A data frame.}

\item{class}{An optional class to give the object.}
}
\description{
Constructor for split objects
}
\keyword{internal}
