% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc_cv}
\alias{mc_cv}
\title{Monte Carlo Cross-Validation}
\usage{
mc_cv(data, prop = 3/4, times = 25, strata = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{prop}{The proportion of data to be retained for modeling/analysis.}

\item{times}{The number of times to repeat the sampling..}

\item{strata}{A variable that is used to conduct stratified sampling to create the resamples.}

\item{...}{Not currently used.}
}
\value{
An object with classes \code{"mc_cv"} and \code{"rset"}. The elements of the object include a tibble called \code{splits} that contains a column for the data split objects and a column called \code{id} that has a character string with the resample identifier.
}
\description{
Monte Carlo Cross-Validation
}
\examples{
mc_cv(mtcars, times = 2)
mc_cv(mtcars, prop = .5, times = 2)
}
