% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc_cv}
\alias{mc_cv}
\title{Monte Carlo Cross-Validation}
\usage{
mc_cv(data, prop = 3/4, times = 25, strata = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{prop}{The proportion of data to be retained for modeling/analysis.}

\item{times}{The number of times to repeat the sampling..}

\item{strata}{A variable that is used to conduct stratified sampling to create the resamples.  (not working yet)}

\item{...}{Not currently used.}
}
\value{
An object with classes \code{"mc_cv"} and \code{"rset"}. The elements of the object include a tibble called \code{splits} that contains a column for the data split objects and a column called \code{id} that has a character string with the resample identifier.
}
\description{
One resample of Monte Carlo cross-validation takes a random sample (without replacement) of the original data set to be used for analysis. All other data points are added to the assessment set.
}
\details{
The \code{strata} argument causes the random sampling to be conducted \emph{within the stratification variable}. The can help ensure that the number of data points in the analysis data is equivalent to the proportions in the original data set.
}
\examples{
mc_cv(mtcars, times = 2)
mc_cv(mtcars, prop = .5, times = 2)
}
