% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{perc_all}
\alias{perc_all}
\title{Percentile wrapper for multiple statistics}
\usage{
perc_all(object, ..., alpha = 0.05)
}
\arguments{
\item{object}{bootstrap resamples created by the \code{bootstraps} function}

\item{...}{parameters to pass to the specific confidence interval methods}

\item{alpha}{level of significance}
}
\description{
Calculate bootstrap confidence interval with percentile method
}
