% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{pretty.vfold_cv}
\alias{pretty.vfold_cv}
\alias{pretty.loo_cv}
\alias{pretty.apparent}
\alias{pretty.rolling_origin}
\alias{pretty.sliding_window}
\alias{pretty.sliding_index}
\alias{pretty.sliding_period}
\alias{pretty.mc_cv}
\alias{pretty.validation_split}
\alias{pretty.nested_cv}
\alias{pretty.bootstraps}
\alias{pretty.permutations}
\alias{pretty.group_vfold_cv}
\alias{pretty.manual_rset}
\title{Short Descriptions of rsets}
\usage{
\method{pretty}{vfold_cv}(x, ...)

\method{pretty}{loo_cv}(x, ...)

\method{pretty}{apparent}(x, ...)

\method{pretty}{rolling_origin}(x, ...)

\method{pretty}{sliding_window}(x, ...)

\method{pretty}{sliding_index}(x, ...)

\method{pretty}{sliding_period}(x, ...)

\method{pretty}{mc_cv}(x, ...)

\method{pretty}{validation_split}(x, ...)

\method{pretty}{nested_cv}(x, ...)

\method{pretty}{bootstraps}(x, ...)

\method{pretty}{permutations}(x, ...)

\method{pretty}{group_vfold_cv}(x, ...)

\method{pretty}{manual_rset}(x, ...)
}
\arguments{
\item{x}{An \code{rset} object}

\item{...}{Not currently used.}
}
\value{
A character vector.
}
\description{
Produce a character vector describing the resampling method.
}
\keyword{internal}
