% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dplyr.R
\name{rsample-dplyr}
\alias{rsample-dplyr}
\title{Compatibility with dplyr}
\description{
rsample should be fully compatible with dplyr 1.0.0.

With older versions of dplyr, there is partial support for the following
verbs: \code{mutate()}, \code{arrange()}, \code{filter()}, \code{rename()}, \code{select()}, and
\code{slice()}. We strongly recommend updating to dplyr 1.0.0 if possible to
get more complete integration with dplyr.
}
