% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample-package.R
\docType{package}
\name{rsample-package}
\alias{rsample}
\alias{rsample-package}
\title{rsample: General Resampling Infrastructure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Classes and functions to create and summarize different types of resampling objects (e.g. bootstrap, cross-validation).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rsample.tidymodels.org}
  \item \url{https://github.com/tidymodels/rsample}
  \item Report bugs at \url{https://github.com/tidymodels/rsample/issues}
}

}
\author{
\strong{Maintainer}: Hannah Frick \email{hannah@rstudio.com} (\href{https://orcid.org/0000-0002-6049-5258}{ORCID})

Authors:
\itemize{
  \item Fanny Chow \email{fannybchow@gmail.com}
  \item Max Kuhn \email{max@rstudio.com}
  \item Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})
  \item Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
