% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_dat}
\alias{two_class_dat}
\title{Sample Two Class Data}
\value{
\item{two_class_dat}{a data frame}
}
\description{
Sample Two Class Data
}
\details{
There are artifical data with two predictors (`A` and `B`) and
 a factor outcome variable (`Class`).
}
\examples{
data(two_class_dat)
str(two_class_dat)
}
\keyword{datasets}
